Make client forget science data after awhile

+++ snis.h
@@ -113,7 +113,7 @@ typedef void (*move_function)(struct snis_entity *o);
 
 struct snis_entity_science_data {
 	char name[20];
-	uint8_t science_data_known;
+	uint16_t science_data_known;
 	uint8_t science_data_requested;
 	uint8_t subclass;
 };
